Complete Phase 2: Extract apply prompts and clean up main prompt module - All 832 tests passing, linter errors resolved
